Top GO enrichment result in the CS-CORE vignette

This data set stores the top GO enrichment result from the analysis in CS-CORE vignette. It is saved for efficiently rendering the vignette without re-evaluating the codes.

Usage

top_enrich_go

Format

top_enrich_go

A list with five elements, where each element holds the top 3 GO terms enriched in a gene module with strong enrichment signals